+++ src/entity.test.ts
@@ -93,6 +93,7 @@ describe('TestEntity', () => {
     protocolVersion: 5,
     username: '',
     password: '',
+    clientId: '',
     ca: '',
     rejectUnauthorized: true,
     cert: '',